Pass down dishId instead of dish from pages/dish
