+++ components/RecipeModal.tsx
@@ -1,14 +1,34 @@
 import { Button, Modal } from "antd";
+import axios from "axios";
+import { useEffect, useState } from "react";
 import RecipeInstructions from "./RecipeInstructions";
 
 type Props = {
   isModalOpen: boolean;
   setIsModalOpen: (value: boolean) => void;
-  dish: Dish | null;
+  dishId: string | null;
 };
 
 const RecipeModal: React.FC<Props> = props => {
-  const { isModalOpen, setIsModalOpen, dish } = props;
+  const { isModalOpen, setIsModalOpen, dishId } = props;
+  const [dish, setDish] = useState(null);
+
+  useEffect(() => {
+    const getDish = async (id: string) => {
+      const result = await axios.get(`/api/dish/${id}`);
+      setDish(result.data.data);
+    };
+
+    // TODO handle error
+    if (dishId) {
+      getDish(dishId).catch(error => console.log(error));
+    } else {
+      setDish(null);
+    }
+    return () => {
+      setDish(null);
+    };
+  }, [dishId]);
 
   const title = dish?.name ? `${dish.name} Recipe` : "Loading...";
   const ingredients = dish?.recipe?.ingredientsText || null;
